+++ napari/_qt/__init__.py
@@ -1,5 +1,5 @@
-from .widgets.qt_range_slider import QHRangeSlider, QVRangeSlider
-from .event_loop import gui_qt
 from ..resources import import_resources
+from .event_loop import gui_qt
+from .widgets.qt_range_slider import QHRangeSlider, QVRangeSlider
 
 import_resources()
