+++ napari/_qt/_constants.py
@@ -1,5 +1,6 @@
-from ..utils.misc import StringEnum
 from enum import auto
+
+from ..utils.misc import StringEnum
 
 
 class LoopMode(StringEnum):
