+++ napari/_qt/layers/qt_image_controls_base.py
@@ -1,14 +1,14 @@
-from functools import partial
 from contextlib import suppress
+from functools import partial
 
 import numpy as np
 from qtpy.QtCore import Qt
 from qtpy.QtGui import QImage, QPixmap
-from qtpy.QtWidgets import QComboBox, QLabel, QSlider, QPushButton
+from qtpy.QtWidgets import QComboBox, QLabel, QPushButton, QSlider
 
+from ..utils import qt_signals_blocked
 from ..widgets.qt_range_slider import QHRangeSlider
 from ..widgets.qt_range_slider_popup import QRangeSliderPopup
-from ..utils import qt_signals_blocked
 from .qt_layer_controls_base import QtLayerControls
