+++ napari/_qt/layers/qt_layer_controls_base.py
@@ -1,5 +1,5 @@
 from qtpy.QtCore import Qt
-from qtpy.QtWidgets import QSlider, QGridLayout, QFrame, QComboBox
+from qtpy.QtWidgets import QComboBox, QFrame, QGridLayout, QSlider
 
 from ...layers.base._base_constants import Blending
