+++ napari/_qt/layers/qt_points_controls.py
@@ -10,11 +10,10 @@ from qtpy.QtWidgets import (
 )
 
 from ...layers.points._points_constants import Mode, Symbol
+from ..utils import disable_with_opacity, qt_signals_blocked
 from ..widgets.qt_color_dialog import QColorSwatchEdit
 from ..widgets.qt_mode_buttons import QtModePushButton, QtModeRadioButton
-from ..utils import qt_signals_blocked
 from .qt_layer_controls_base import QtLayerControls
-from ..utils import disable_with_opacity
 
 
 class QtPointsControls(QtLayerControls):
