+++ napari/_qt/layers/qt_surface_controls.py
@@ -1,5 +1,6 @@
+from qtpy.QtWidgets import QHBoxLayout, QLabel
+
 from .qt_image_controls_base import QtBaseImageControls
-from qtpy.QtWidgets import QLabel, QHBoxLayout
 
 
 class QtSurfaceControls(QtBaseImageControls):
